Decompile func0f0cf728

+++ src/include/game/data/data_0160b0.h
@@ -95,7 +95,7 @@ extern u32 var80070e90;
 extern u32 var80070e9c;
 extern u32 var80070ea8;
 extern u32 var80070ec0;
-extern u32 var80070ec4;
+extern u8 var80070ec4;
 extern f32 g_EyespyMaxHeight;
 extern u32 var80070ed4;
 extern u32 var80070ed8;
